cloudhand-api/pyproject.toml
[project]
name = "cloudhand-control-plane"
version = "0.1.0"
description = "Cloudhand Control Plane (FastAPI + Cloudhand CLI)"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-multipart",
    "requests",
    "httpx",
    "SQLAlchemy",
    "asyncpg",
    "alembic",
    "PyJWT",
    "cryptography",
    "hvac",
    "paramiko",
    "click",
    "e2b",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
